pyproject.toml
[tool.poetry]
name = "bayesian-stats"
version = "0.1.0"
description = "Sanbox library for learning Bayesian statistics and probabilistic programming"
authors = ["libertininick <libertini.nick@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "bayesian_stats", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
conda-lock = "^2.1.0"
numpy = "^1.25.0"
arviz = "^0.15.1"
matplotlib = "^3.7.1"
pandas = "^2.0.2"
scipy = "^1.11.0"
xarray = "^2023.6.0"
pyro-ppl = "^1.8.6"
torch = ">=2.0.0, !=2.0.1, !=2.1.0"
python-dotenv = "^1.0.0"
einops = "^0.7.0"


[tool.poetry.group.dev.dependencies]
coverage = "^7.2.7"
mypy = "^1.4.1"
typeguard = "^4.0.0"
pytest = "^7.4.0"
hypothesis = "^6.88.1"
pytest-check = "^2.2.2"
ruff = "^0.1.6"


[tool.poetry.group.jupyter.dependencies]
ipykernel = "^6.25.1"
ipympl = "^0.9.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.coverage.run]
branch = true
data_file = "build/.coverage"
source = [
    "src",
    "tests",
]

[tool.coverage.report]
exclude_also = [
    "def __repr__"
]
fail_under = 90
precision = 1
show_missing = true
skip_covered = true
skip_empty = true


[tool.mypy]
cache_dir = "./build/mypy"
disallow_untyped_defs = true
ignore_missing_imports = true
plugins = "numpy.typing.mypy_plugin"
pretty = true
show_column_numbers = true
show_error_context = true
warn_unreachable = true


[tool.pytest.ini_options]
addopts = "-rf --doctest-modules --strict-markers"
cache_dir = "./build/pytest"
console_output_style = "progress"
markers = [
    "slow"
]
minversion = "6.0"
testpaths = [
    "src",
    "tests",
]


[tool.ruff]
cache-dir = "./build/ruff"
line-length = 79  # set max line length 
indent-width = 4
target-version = "py311"  # python version of code

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
pydocstyle = {convention = "numpy"}
mccabe = {max-complexity = 10}
select = [
    "B",  # flake8-bugbear
    "D",  # pydocstyle
    "E",  # pycodestyle
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "S",  # flake8-bandit
    "C4",  # flake8-comprehensions
    "C9",  # max-complexity
    "UP",  # pyupgrade
    "ANN",  # flake8-annotations
    "RUF",  # Ruff-specific rules
    "SIM",  # flake8-simplify
    "E501",   # `line-too-long` rule
]
ignore = [
    "ANN002",  # Missing type annotation for `*args`
    "ANN003",  # Missing type annotation for `**kwargs`
    "ANN101",  # missing-type-self
    "ANN102",  # missing-type-cls
    "ANN401",  # disallow any-type
]
per-file-ignores = {"**/{tests}/*" = ["S101"]}  # ignore asserts in tests